+++ gen-protos.sh
@@ -17,7 +17,7 @@ SHARED_PROTO_PATH=$SHARED_SERVER_DIR/proto
 
 # define the path to the payment-server directory
 PAYMENT_SERVER_DIR=payment-server
-PAYMENT_PROTO_PATH=$PAYMENT_SERVER_DIR/proto
+#PAYMENT_PROTO_PATH=$PAYMENT_SERVER_DIR/proto
 
 # define the path to the mobile directory
 MOBILE_DIR=mobile
@@ -42,23 +42,19 @@ protoc -I=$MEDIA_PROTO_PATH --go_out=$MEDIA_SERVER_DIR/gen --go_opt=paths=source
   --go-grpc_out=$MEDIA_SERVER_DIR/gen --go-grpc_opt=paths=source_relative \
   $(find $MEDIA_PROTO_PATH -iname "*.proto")
 
-# generate for payment-server ruby server into gen folder
-grpc_tools_ruby_protoc -I=$PAYMENT_PROTO_PATH --ruby_out=$PAYMENT_SERVER_DIR/gen  --grpc_out=$PAYMENT_SERVER_DIR/gen \
-  $(find $PAYMENT_PROTO_PATH -iname "*.proto")
-
 # generate for shared-server using python
 protoc -I=$SHARED_PROTO_PATH --go_out=$SHARED_SERVER_DIR/gen --go_opt=paths=source_relative \
   --go-grpc_out=$SHARED_SERVER_DIR/gen --go-grpc_opt=paths=source_relative \
   $(find $SHARED_PROTO_PATH -iname "*.proto")
 
 # copy the proto file from media-server to auth-server
-cp $MEDIA_PROTO_PATH/media.proto $AUTH_PROTO_PATH/media.proto
+#cp $MEDIA_PROTO_PATH/media.proto $AUTH_PROTO_PATH/media.proto
 
 # generate for auth-server using rust
-cargo build --manifest-path $AUTH_SERVER_DIR/Cargo.toml
+#cargo build --manifest-path $AUTH_SERVER_DIR/Cargo.toml
 
 # generate for flutter using dart
 protoc -I=$AUTH_PROTO_PATH -I=$SHARED_PROTO_PATH -I=$CORE_PROTO_PATH \
   --dart_out=grpc:$MOBILE_OUT_DIR \
-  $(find $SHARED_PROTO_PATH -iname "*.proto")
+  $(find $CORE_PROTO_PATH -iname "*.proto")
 #  $(find $AUTH_PROTO_PATH -iname "*.proto") $(find $SHARED_PROTO_PATH -iname "*.proto") $(find $CORE_PROTO_PATH -iname "*.proto")