+++ mobile/lib/core/network/token.interceptor.dart
@@ -3,7 +3,6 @@ import 'dart:async';
 import 'package:grpc/grpc.dart';
 import 'package:injectable/injectable.dart';
 import 'package:mobile/core/utils/session.dart';
-import 'package:shared_utils/shared_utils.dart';
 
 @injectable
 class TokenGrpcInterceptor implements ClientInterceptor {
@@ -21,9 +20,6 @@ class TokenGrpcInterceptor implements ClientInterceptor {
         },
       ),
     );
-
-    logger.d(
-        'grpc streaming method: ${method.path} => metadata: ${newOpts.metadata}');
     // do nothing for now
     return invoker(method, requests, newOpts);
   }
@@ -39,9 +35,6 @@ class TokenGrpcInterceptor implements ClientInterceptor {
         },
       ),
     );
-    logger.d(
-        'grpc unary method: ${method.path} => metadata: ${newOpts.metadata} => payload : $self');
-    return invoker(method, self, newOpts)
-      ..then((data) => logger.d('grpc unary response => $data'));
+    return invoker(method, self, newOpts);
   }
 }