+++ mobile/lib/core/utils/validator.dart
@@ -1,3 +1,5 @@
+import 'package:flutter/material.dart';
+import 'package:mobile/core/utils/extensions.dart';
 import 'package:shared_utils/shared_utils.dart' show TextX;
 
 class Validators {
@@ -7,34 +9,38 @@ class Validators {
     return null;
   }
 
-  static String? validateAuthCode(String? input) {
+  static String? validateAuthCode(BuildContext context, String? input) {
     input = input?.replaceAll(' ', '');
-    if (input.isNullOrEmpty()) return 'Required';
+    if (input.isNullOrEmpty()) return context.localizer.required;
     var regex = RegExp(r'^\d{6}$');
-    return regex.hasMatch(input!.trim()) ? null : 'Auth code too short';
+    return regex.hasMatch(input!.trim()) ? null : context.localizer.invalidAuthCode;
   }
 
-  static String? validatePhone(String? input) {
+  static String? validatePhone(BuildContext context, String? input) {
     input = input?.replaceAll(' ', '');
-    if (input.isNullOrEmpty()) return 'Required';
+    if (input.isNullOrEmpty()) return context.localizer.required;
     var regex =
         RegExp(r'^[\+]?[(]?[0-9]{3}[)]?[-\s\.]?[0-9]{3}[-\s\.]?[0-9]{4,6}$');
-    return regex.hasMatch(input!.trim()) ? null : 'Invalid phone number';
+    return regex.hasMatch(input!.trim()) ? null : context.localizer.invalidPhoneNumber;
   }
 
-  static String? validateEmail(String? input) {
-    if (input.isNullOrEmpty()) return 'Required';
+  static String? validateEmail(BuildContext context, String? input) {
+    if (input.isNullOrEmpty()) return context.localizer.required;
     var regex = RegExp(r'(^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$)');
-    return regex.hasMatch(input!.trim()) ? null : 'Invalid email address';
+    return regex.hasMatch(input!.trim()) ? null : context.localizer.invalidEmail;
   }
 
-  static String? validatePassword(String? input) {
-    if (input.isNullOrEmpty()) return 'Required';
+  static String? validatePassword(BuildContext context, String? input,
+      [String? confirmPassword]) {
+    if (input.isNullOrEmpty()) return context.localizer.required;
 
     /// min length of 4
     var regex = RegExp(r'[A-Za-z0-9@#$%^&+=]{4,}');
-    return regex.hasMatch(input!.trim())
-        ? null
-        : 'Invalid password or too short';
+    var validated = regex.hasMatch(input!.trim());
+    if (!validated) return context.localizer.passwordTooShort;
+    if (confirmPassword != null && input != confirmPassword) {
+      return context.localizer.passwordsDoNotMatch;
+    }
+    return null;
   }
 }